# src/config/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "ElegancceAPI",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/reservations": {
      "get": {
        "tags": [
          "reservations"
        ],
        "summary": "Busca todas as reservas e seus procedimentos",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "reservations"
        ],
        "summary": "Cria uma nova reserva",
        "description": "",
        "parameters": [
          {
            "name": "reservation",
            "in": "body",
            "decription": "Modelo de dados a ser seguido",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "number",
                  "example": 4
                },
                "date": {
                  "type": "number",
                  "example": 1658838525051
                },
                "procedures": {
                  "type": "array",
                  "example": [
                    1,
                    2,
                    3,
                    4
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "user_id",
                "date",
                "procedures"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "reservations"
        ],
        "summary": "Atualiza uma reserva",
        "description": "",
        "parameters": [
          {
            "name": "reservation",
            "in": "body",
            "decription": "Modelo de dados a ser seguido",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 4
                },
                "user_id": {
                  "type": "number",
                  "example": 4
                },
                "date": {
                  "type": "number",
                  "example": 1658838525051
                },
                "procedures": {
                  "type": "array",
                  "example": [
                    1,
                    2,
                    3,
                    4
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "id",
                "user_id",
                "date",
                "procedures"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reservations/user/{id}": {
      "get": {
        "tags": [
          "reservations"
        ],
        "summary": "Busca reservas relaciondas a um usuario especifico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/reservations/{id}": {
      "delete": {
        "tags": [
          "reservations"
        ],
        "summary": "Deleta uma reserva",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/requests": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Busca todos os pedidos",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Realiza um novo pedido",
        "description": "",
        "parameters": [
          {
            "name": "",
            "in": "body",
            "description": "Modelo de dados",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "number",
                  "example": 1
                },
                "products": {
                  "type": "object",
                  "properties": {
                    "qt_product": {
                      "type": "number",
                      "example": 22
                    },
                    "product_id": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "required": [
                    "qt_product",
                    "product_id"
                  ]
                }
              },
              "required": [
                "user_id",
                "products"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "requests"
        ],
        "summary": "Atualiza um novo pedido",
        "description": "",
        "parameters": [
          {
            "name": "",
            "in": "body",
            "description": "Modelo de dados",
            "schema": {
              "type": "object",
              "properties": {
                "request_id": {
                  "type": "number",
                  "example": 1
                },
                "user_id": {
                  "type": "number",
                  "example": 1
                },
                "products": {
                  "type": "object",
                  "properties": {
                    "qt_product": {
                      "type": "number",
                      "example": 22
                    },
                    "product_id": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "required": [
                    "qt_product",
                    "product_id"
                  ]
                }
              },
              "required": [
                "request_id",
                "user_id",
                "products"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/requests/user/{id}": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Busca todos os pedidos",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/requests/{id}": {
      "delete": {
        "tags": [
          "requests"
        ],
        "summary": "Delete um pedido",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}